moderation: Clarify warning ID command args. (#1800)
* moderation: Use "WarningId" arg. name for edits

People are persistently confused about the type of ID
they should be using, when interacting with warnings via command.
This commit updates the `"EditWarning"` command to use the arg. name,
`"WarningId"` to hopefully clarify this.

* moderation: Use "WarningId" arg. name for deleting

People are persistently confused about the type of ID
they should be using, when interacting with warnings via command.
This commit updates the `"DelWarning"` command to use the arg. name,
`"WarningId"`, to hopefully clarify this.