cc/cron: fix blacklist issues (#1799)
* cc/cron: add thorough tests

Signed-off-by: SoggySaussages <vmdmaharaj@gmail.com>

* cc/cron: distinguish between all set and *

Currently, cron parse parses a * in hours the same as 0,1,2...,23, and similar for dow. This poses an issue since
0,1,2...,23 != *. Now if all hours/dows are set, the parse will leave the input unchanged rather than forcing "*"s to
"0,1,2...23"s.

Signed-off-by: SoggySaussages <vmdmaharaj@gmail.com>

* cc/cron: stop creating undesired OR for days expressions

if month/dom and dow are both configured, cron will process them using OR. If user didn't want that, they wouldn't set
both in the expression. but if using day of week blacklist, in the way we do it currently that would configure it, and
then cron would use OR and mess up the user's schedule if they explicitly avoided this behaviour in their expression.
For this reason, this commit now only uses the existing method if month/dom aren't configured, or if week is. Otherwise
it will loop recalculation of cron schedule until a new time is found which satisfies the dow blacklist.

Signed-off-by: SoggySaussages <vmdmaharaj@gmail.com>

* cc/cron: bump to beta

Signed-off-by: SoggySaussages <vmdmaharaj@gmail.com>

* cc/cron: remove debug logging

Signed-off-by: SoggySaussages <vmdmaharaj@gmail.com>

* cc:

* cc/cron: hard limit on next time loop

Limit the edge case backup blacklist checker to 200 iterations

* cc/interval: use table-driven tests

* Correct test output

* Ensure test crons are valid

* Make test reference times more consistent

* Correct exclude days data

---------

Signed-off-by: SoggySaussages <vmdmaharaj@gmail.com>
Co-authored-by: Joseph Liu <jo3.l.dev@outlook.com>